Chatbot/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.10.3
python-dotenv==1.0.1
mysql-connector-python==9.1.0
requests==2.32.3
